+++ app/change-password/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import Button from "@/components/button";
+import Button from "@/components/global/button";
 import Input from "@/components/global/input";
 
 import Image from "next/image";
